pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-project-template"
version = "0.1.0"
description = "A modern Django project template"
readme = "README.md"
authors = [
    {name = "Valor Engels", email = "your_email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
requires-python = ">=3.8"
dependencies = [
    "Django>=4.0",
    "django-extensions",
    "django-filter",
    "django-debug-toolbar",
    "django-cors-headers",
    "djangorestframework",
    "djangorestframework-api-key",
    "psycopg2",
    "python-dotenv",
    "dj-database-url",
    "gunicorn",
    "whitenoise",
    "django-storages",
    "django-widget-tweaks",
    "django-htmx",
    "django-unfold",
]

[project.optional-dependencies]
dev = [
    "django-browser-reload",
    "Werkzeug",
    "aider-chat",
    "black",
    "flake8",
    "mypy",
    "django-stubs",
    "ipython",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-django",
    "factory-boy",
]

[project.urls]
Homepage = "https://github.com/yourusername/django-project-template"
Issues = "https://github.com/yourusername/django-project-template/issues"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.django-stubs]
django_settings_module = "settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings"
python_files = "test_*.py"
testpaths = ["apps"]